requirements.txt
urllib3
requests
lxml
python-dateutil
pyOpenSSL
